Add 2 different AIs to game and a difficulty setting
Added two different AIs based on the position value idea. The AI3 function incorporates the AI1 function to choose between the top moves with similar values. No idea how effective this is - it still needs to be checked.

Will add the source for the move value theory at a later date.