Stop AI3 passing AI1 an empty list
In the previous implementation of AI3 it called AI1 on its own refined
list of moves. However, there was no check to make sure there were any
moves in that list and so in the relatively uncommon case that there were
no moves with the top 5 values, an error would occur, rather than
continuing on to the lower valued move positions.

See lines from 575 - specifically line 580.