Create global var to stop out-of-turn clicks
Previously, if a user clicked repeatedly after their move the clicks
would become back logged, and the program would get overloaded. This
behaviour could also result in a move being drawn based on an
out-of-date gamestate, or even just a user getting two turns in a row.

To prevent this, a global moveInProgress has been created that is set to
True at each players turn, and a game board click only triggers userMove
if the moveInProgress variable is set to False.