Fix error with newGame(), clean drawLoadedPieces()
There was a bug when a game had ended and newGame was called. This was due
to the global gameHasEnded getting reset to False after the newGame()
function was called. This has been fixed. Additionally, I noticed that the
drawLoadedPieces() function could be simplified by just using the
pre-existing drawPieces() function. This has been updated.

For masterReversi.py, these changes were implemented in addition to
allowing the loadGame to work with the new image stamps.