Complete revision of constants and board layout
In accordance with Shunyi's feedback, I reverted the board to the 0,0
centre coordinates. This is the only way the game will stay centered
when shrunk.

I also added wholesale changes to the constants. These include for the
buttons, for the large and small popups. Revisions will have to be made
in order to optimize the exact use of those constants, as some are
repeated but I think the general concept is what our TA wants.

I also added text to the rules/instructions.