Refine display, add world coord option
Made small edits to where the title and buttons are drawn in order to
bring everything in closer. Also added a commented out section that
features valid world coordinate settings that can be used if desired.
However these won't work with the font sizes, as they created in
reference to PIECE_SIZE. Also, I don't think they want us to use this
feature.