Add file save error checking, new button actions
Improved functionality of save game - file overwritten if confirmed. This
avoids having to delete the file on load, for our demo there won't be a
problem with multiple files - we can always manually delete.

Added optional argument for newGame() function to allow a direct new game,
or a direct save, or give the user the option. If they cancel on load
game, it displays an error while looping the prompt.

There's also now a super secret feature - change the difficulty settings!