Refine hints functions, update buttons
Hints are now shown only for the human player, and are toggled on/off by
the old difficulty button. They are triggered in two places: the first
is when the newGame function is called, and the first player is the
human player. The second place is at the end of the computerMove when
the turn switches back to the human player.