Allow game difficulty and hints status to be saved
Now difficulty level and hints status is saved to the users game file. If
it is an old game file with no hints or diff info, the diff will just be
easy and the hints will be enabled. All games from now on will save this
info though.

Still need a good way of indicating game difficulty.